Continuations: Clean up return_path_matches logic
The previous "URI unescape until we can't anymore, then match" logic
was necessary because different server request methods had different
standards of escaping.  PSGI has a clear standard, so we can do the
much simpler check of if the paths match, the only argument is
J:RETURN, and it is set correctly.